edit kill20.sh

+++ Lesson7/kill20.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
 
-id=$(ps -aux | grep 'apache2 -k start' | awk '$1=$1' | cut -d " " -f2 | head -n 1)
+id=$(ps -aux | grep '[a]pache2 -k start' | awk '$1=$1' | cut -d " " -f2 | head -n 1 | tr -d '/n')
 
-name=$(ps -aux | grep 'apache2 -k start' | awk '$1=$1' | cut -d " " -f11 | head -n 1)
+name=$(ps -aux | grep '[a]pache2 -k start' | awk '$1=$1' | cut -d " " -f11 | head -n 1)
 
-echo "$id"
+if [ -z $id ]; then
+  
+  echo "Process not found"
 
-`kill -20 $id`
+else
+	
+  `kill -20 $id`
 
-echo "process $name $id suspended"
+  echo "process $name $id suspended"
 
-sleep 30
+  sleep 30
+
+fi
+
